[#26] GroupedTimeSlot 분리, TimeSlot에 generate함수 추가
