+++ TIG/Sources/Utility/Extension/Array+Ext.swift
@@ -48,3 +48,16 @@ extension Array where Element == TimeSlot {
     return result
   }
 }
+
+
+extension Array where Element == GroupedTimeSlot {
+  /// 배열에서 현재 시간대에 위치한 TimeSlot 객체를 반환합니다. (빈 배열인 경우 mock 반환)
+  var currentTimeSlot: GroupedTimeSlot {
+    let now = Date()
+    let totalSeconds = now.totalSeconds
+    
+    return self.first(where: {
+      $0.start <= totalSeconds && $0.end > totalSeconds
+    }) ?? .mock
+  }
+}
